Move the gui part of the settings into gui/ and refactor the way settings are accessed to make it simpler.
